terminé sans bonus

+++ OFAQ/src/Controller/backend/TagController.php
@@ -15,19 +15,15 @@ use Symfony\Component\Routing\Annotation\Route;
  */
 class TagController extends Controller
 {
-    /**
-     * @Route("/", name="tag_index", methods="GET")
-     */
-    public function index(TagRepository $tagRepository): Response
-    {
-        return $this->render('tag/index.html.twig', ['tags' => $tagRepository->findAll()]);
-    }
+    
 
     /**
      * @Route("/new", name="tag_new", methods="GET|POST")
      */
     public function new(Request $request): Response
     {
+        $this->denyAccessUnlessGranted('ROLE_MODERATOR', null, 'Impossible d\'accéder à cette page!');
+
         $tag = new Tag();
         $form = $this->createForm(TagType::class, $tag);
         $form->handleRequest($request);
@@ -49,16 +45,17 @@ class TagController extends Controller
     /**
      * @Route("/{id}", name="tag_show", methods="GET")
      */
-    public function show(Tag $tag): Response
+   /* public function show(Tag $tag): Response
     {
         return $this->render('tag/show.html.twig', ['tag' => $tag]);
-    }
+    }*/
 
     /**
      * @Route("/{id}/edit", name="tag_edit", methods="GET|POST")
      */
     public function edit(Request $request, Tag $tag): Response
     {
+        $this->denyAccessUnlessGranted('ROLE_MODERATOR', null, 'Impossible d\'accéder à cette page!');
         $form = $this->createForm(TagType::class, $tag);
         $form->handleRequest($request);
 
@@ -79,6 +76,7 @@ class TagController extends Controller
      */
     public function delete(Request $request, Tag $tag): Response
     {
+        $this->denyAccessUnlessGranted('ROLE_MODERATOR', null, 'Impossible d\'accéder à cette page!');
         if ($this->isCsrfTokenValid('delete'.$tag->getId(), $request->request->get('_token'))) {
             $em = $this->getDoctrine()->getManager();
             $em->remove($tag);
@@ -86,5 +84,21 @@ class TagController extends Controller
         }
 
         return $this->redirectToRoute('tag_index');
+    }
+
+
+        //Affiche les questions affiliées à un tag 
+
+    /**
+     * @Route("/{id}/questions", name="tag_questions", methods="GET")
+     */
+    public function questionList(Tag $tag) {
+
+        $questions = $tag->getQuestions();
+
+       return $this->render('question/index.html.twig', [
+            'questions'=>$questions,
+        ] );
+
     }
 }
